+++ __tests__/bank-account.test.ts
@@ -38,7 +38,7 @@ describe("[UNIT] => BANK ACCOUNTS ROUTER", () => {
       .set("Authorization", `Bearer ${playerAccess} `);
 
     expect(response.status).toBe(BAD_REQUEST);
-    expect(response.body.details[0].type).toBe("unknown_fields");
+    expect(response.body.data[0].type).toBe("unknown_fields");
   });
 
   it.each`
@@ -62,8 +62,8 @@ describe("[UNIT] => BANK ACCOUNTS ROUTER", () => {
         .set("Authorization", `Bearer ${playerAccess} `);
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.details[0].msg).toBe(message);
-      expect(response.body.details[0].path).toBe(field);
+      expect(response.body.data[0].msg).toBe(message);
+      expect(response.body.data[0].path).toBe(field);
     },
   );