+++ __tests__/transactions.test.ts
@@ -68,8 +68,8 @@ describe("[UNIT] => TRANSACTIONS", () => {
         .set("User-Agent", USER_AGENT);
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.details[0].msg).toBe(message);
-      expect(response.body.details[0].path).toBe(field);
+      expect(response.body.data[0].msg).toBe(message);
+      expect(response.body.data[0].path).toBe(field);
     });
 
     it("Should return 400 unknown_field", async () => {
@@ -83,7 +83,7 @@ describe("[UNIT] => TRANSACTIONS", () => {
         .set("User-Agent", USER_AGENT);
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.details[0].type).toBe("unknown_fields");
+      expect(response.body.data[0].type).toBe("unknown_fields");
     });
 
     it("Should return 401", async () => {
@@ -275,8 +275,8 @@ describe("[UNIT] => TRANSACTIONS", () => {
         .set("User-Agent", USER_AGENT);
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.details[0].msg).toBe(message);
-      expect(response.body.details[0].path).toBe(field);
+      expect(response.body.data[0].msg).toBe(message);
+      expect(response.body.data[0].path).toBe(field);
     });
 
     it("Should return 400 unknown_field", async () => {
@@ -290,7 +290,7 @@ describe("[UNIT] => TRANSACTIONS", () => {
         .set("User-Agent", USER_AGENT);
 
       expect(response.status).toBe(BAD_REQUEST);
-      expect(response.body.details[0].type).toBe("unknown_fields");
+      expect(response.body.data[0].type).toBe("unknown_fields");
     });
 
     it("Should return 401", async () => {