+++ src/components/players/services.ts
@@ -9,9 +9,9 @@ import {
 import { LoginResponse, PlainPlayerResponse } from "@/types/response/players";
 import { compare, hash } from "@/utils/crypt";
 import { hidePassword } from "@/utils/auth";
-import { CustomError } from "@/middlewares/errorHandler";
 import { HttpService } from "@/services/http.service";
 import { ERR } from "@/config/errors";
+import { CustomError } from "@/helpers/error/CustomError";
 
 export class PlayerServices {
   /**