+++ src/helpers/error/ForbiddenError.ts
@@ -1,14 +1,10 @@
-import { FORBIDDEN } from "http-status/lib";
+import httpStatus, { FORBIDDEN } from "http-status/lib";
+import { CustomError } from "./CustomError";
 
-class ForbiddenError {
-  readonly status: number;
-  readonly message: string;
-
-  constructor(message: string) {
+class ForbiddenError extends CustomError {
+  constructor(description: string) {
+    super({ status: FORBIDDEN, code: httpStatus[FORBIDDEN], description });
     Object.setPrototypeOf(this, new.target.prototype);
-
-    this.status = FORBIDDEN;
-    this.message = message;
   }
 }