+++ src/middlewares/rate-limiters/deposit.ts
@@ -1,8 +1,8 @@
 import { RateLimiterMemory } from "rate-limiter-flexible";
 import { DepositRequest } from "@/types/request/transfers";
-import { CustomError } from "@/middlewares/errorHandler";
 import { ERR } from "@/config/errors";
 import CONFIG from "@/config";
+import { CustomError } from "@/helpers/error/CustomError";
 
 const rateLimiter = new RateLimiterMemory({
   points: CONFIG.APP.ENV === CONFIG.SD.ENVIRONMENTS.TEST ? 10 : 1,