+++ src/types/response/error.ts
@@ -1,6 +1,8 @@
-export interface ErrorData {
+import { ValidationError } from "express-validator";
+
+export type ErrorData = {
   status: number;
   code: string;
-  description: string;
+  description: string | ValidationError[];
   detail?: object;
-}
+};